Refactoring of XFInit for VRP and PDP
