+++ src/main/java/xf/xfvrp/opt/init/precheck/PreCheckException.java
@@ -2,6 +2,8 @@ package xf.xfvrp.opt.init.precheck;
 
 public class PreCheckException extends Exception {
 
+	private static final long serialVersionUID = 1L;
+
 	public PreCheckException(String msg) {
 		super(msg);
 	}
