+++ src/main/java/xf/xfvrp/opt/init/solution/pdp/XFPDPInit.java
@@ -1,0 +1,18 @@
+package xf.xfvrp.opt.init.solution.pdp;
+
+/** 
+ * Copyright (c) 2012-present Holger Schneider
+ * All rights reserved.
+ *
+ * This source code is licensed under the MIT License (MIT) found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ **/
+public class XFPDPInit {
+
+	public static boolean IGNORE_IMPROPER_AMOUNTS = true;
+
+	
+
+	
+}
